+++ protected/humhub/modules/content/widgets/views/wallEntryControls.php
@@ -2,6 +2,7 @@
 
 use humhub\components\View;
 use humhub\helpers\Html;
+use humhub\widgets\menu\DropdownDivider;
 use humhub\widgets\menu\MenuEntry;
 use humhub\widgets\menu\WidgetMenuEntry;
 use humhub\widgets\menu\DropdownMenu;
@@ -26,7 +27,10 @@ use humhub\widgets\menu\DropdownMenu;
 
     <ul class="dropdown-menu dropdown-menu-end">
         <?php foreach ($entries as $entry) : ?>
-            <?php if ($entry instanceof WidgetMenuEntry) : ?>
+            <?php if (!$entry->isVisible()) : ?>
+                <?php continue ?>
+            <?php elseif ($entry instanceof WidgetMenuEntry || $entry instanceof DropdownDivider) : ?>
+                <?php // Both render their own `<li>`. ?>
                 <?= $entry->render() ?>
             <?php else: ?>
                 <li>